Changed water respawn script to respawn plane
