+++ Capstone-Game/Assets/Resources/Prefabs/Terrain/Water.prefab
@@ -1,5 +1,48 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &380148735777636868
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 7856287135317698330}
+  - component: {fileID: 2309324455077116291}
+  m_Layer: 4
+  m_Name: RespawnPlane
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &7856287135317698330
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 380148735777636868}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: -1, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 7589957227752554003}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2309324455077116291
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 380148735777636868}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c57dc9942775f834ba65b0279d5a4508, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &7589957227752554007
 GameObject:
   m_ObjectHideFlags: 0
@@ -12,7 +55,6 @@ GameObject:
   - component: {fileID: 7589957227752554002}
   - component: {fileID: 7589957227752554005}
   - component: {fileID: 7589957227752554004}
-  - component: {fileID: 7698266402831919927}
   m_Layer: 4
   m_Name: Water
   m_TagString: Untagged
@@ -30,7 +72,8 @@ Transform:
   m_LocalRotation: {x: -0, y: -0, z: -4.3786486e-10, w: 1}
   m_LocalPosition: {x: 0, y: -2, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
+  m_Children:
+  - {fileID: 7856287135317698330}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -99,17 +142,3 @@ MonoBehaviour:
   waveFrequency: 0.3
   waveLength: 0.5
   waveOriginPosition: {x: 0, y: 0, z: 0}
---- !u!114 &7698266402831919927
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7589957227752554007}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c57dc9942775f834ba65b0279d5a4508, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  squirrel: {fileID: 0}
-  depth: 0.5
